requirements.txt
# BIS Standards RAG Engine
# Requirements for the BIS Recommendation Engine

# Core dependencies
faiss-cpu==1.7.4
numpy==1.24.3
pandas==2.0.3

# Sentence transformers (includes transformers, torch)
sentence-transformers>=2.2.2

# PDF processing
pypdf==3.9.0
pdfminer.six==20221105

# API and serving
fastapi==0.104.1
uvicorn==0.24.0

# Utilities
tiktoken==0.5.2
tenacity==8.2.3

# For evaluation
scikit-learn==1.3.2

# Document Processing
PyMuPDF==1.23.6
python-docx==1.1.0
python-multipart==0.0.6
reportlab==4.0.9